refactor: remove legacy integration tests
