--- global-playwright-setup.js
@@ -1,10 +1,0 @@
-const dotenv = require('dotenv');
-
-async function globalSetup() {
-  dotenv.config({
-    path: '.env',
-    override: true,
-  });
-}
-
-module.exports = globalSetup;
