+++ package.json
@@ -22,19 +22,15 @@
     "lint:eslint": "eslint \"src/**/*.{ts,tsx}\"",
     "lint:filename": "ls-lint",
     "lint:fix": "eslint \"src/**/*.{ts,tsx}\" --fix",
-    "lint:prettier": "prettier --check \"{src,tests,tests-legacy}/**/*.{ts,tsx}\" \"*.{js,json}\"",
-    "lint:prettier:fix": "prettier --write \"{src,tests,tests-legacy}/**/*.{ts,tsx}\" *.js",
+    "lint:prettier": "prettier --check \"{src,tests}/**/*.{ts,tsx}\" \"*.{js,json}\"",
+    "lint:prettier:fix": "prettier --write \"{src,tests}/**/*.{ts,tsx}\" *.js",
     "lint:unused-exports": "ts-unused-exports tsconfig.json --ignoreFiles=leather-styles",
     "lint:remote-wallet-config": "npx ajv-cli validate -s config/wallet-config.schema.json -d config/wallet-config.json",
     "lint:deps": "dependency-cruise  --config .dependency-cruiser.js \"src/**/*.{ts,tsx}\"",
     "prod:ext": "yarn build",
     "prod:analyze": "cross-env WALLET_ENVIRONMENT=production ANALYZE=true webpack -p",
-    "test:integration": "jest --config=./jest.integration.config.js --verbose=true --runInBand --testPathPattern=./tests-legacy/integration/*",
-    "test:integration:ci": "jest --config=./jest.integration.config.js --testPathPattern=./tests-legacy/integration/*",
-    "test:integration-api": "jest --config=./jest.integration.config.js --verbose=true --runInBand --testPathPattern=./tests-legacy/test-api/*",
-    "test:integration-api:ci": "jest --config=./jest.integration.config.js --testPathPattern=./tests-legacy/test-api/*",
+    "test:integration": "playwright test",
     "test:unit": "vitest run",
-    "test": "WALLET_ENVIRONMENT=testing jest --verbose=true",
     "test:coverage": "WALLET_ENVIRONMENT=testing jest --collect-coverage",
     "test:watch": "WALLET_ENVIRONMENT=testing jest --watch",
     "typecheck": "tsc --noEmit",
@@ -146,22 +142,22 @@
     "@segment/analytics-next": "1.56.0",
     "@sentry/react": "7.72.0",
     "@sentry/tracing": "7.72.0",
-    "@stacks/auth": "6.5.4",
+    "@stacks/auth": "6.9.0",
     "@stacks/blockchain-api-client": "6.3.4",
-    "@stacks/common": "6.5.2",
-    "@stacks/connect": "7.2.1",
-    "@stacks/connect-ui": "6.0.0",
-    "@stacks/encryption": "6.5.4",
-    "@stacks/network": "6.5.4",
-    "@stacks/profile": "6.5.4",
+    "@stacks/common": "6.8.1",
+    "@stacks/connect": "7.4.0",
+    "@stacks/connect-ui": "6.1.1",
+    "@stacks/encryption": "6.9.0",
+    "@stacks/network": "6.8.1",
+    "@stacks/profile": "6.9.0",
     "@stacks/rpc-client": "1.0.3",
-    "@stacks/storage": "6.5.4",
+    "@stacks/storage": "6.9.0",
     "@stacks/transactions": "6.9.0",
     "@stacks/ui": "7.10.0",
     "@stacks/ui-core": "7.3.0",
     "@stacks/ui-theme": "7.5.0",
     "@stacks/ui-utils": "7.5.0",
-    "@stacks/wallet-sdk": "6.5.4",
+    "@stacks/wallet-sdk": "6.9.0",
     "@stitches/react": "1.2.8",
     "@styled-system/theme-get": "5.1.2",
     "@tanstack/query-sync-storage-persister": "4.35.3",
@@ -250,11 +246,9 @@
     "@redux-devtools/remote": "0.8.1",
     "@schemastore/web-manifest": "0.0.6",
     "@sentry/webpack-plugin": "2.4.0",
-    "@stacks/connect-react": "22.1.1",
+    "@stacks/connect-react": "22.2.0",
     "@stacks/prettier-config": "0.0.10",
     "@stacks/stacks-blockchain-api-types": "6.3.4",
-    "@swc-node/jest": "1.5.6",
-    "@swc/core": "1.3.32",
     "@trivago/prettier-plugin-sort-imports": "4.2.0",
     "@types/argon2-browser": "1.18.2",
     "@types/chroma-js": "2.4.1",
@@ -263,8 +257,6 @@
     "@types/download": "8.0.2",
     "@types/expect-puppeteer": "5.0.3",
     "@types/html-webpack-plugin": "3.2.7",
-    "@types/jest": "29.2.6",
-    "@types/jest-dev-server": "5.0.0",
     "@types/jsdom": "21.1.1",
     "@types/lodash.get": "4.4.7",
     "@types/node": "20.7.1",
@@ -310,10 +302,6 @@
     "file-loader": "6.2.0",
     "generate-json-webpack-plugin": "2.0.0",
     "html-webpack-plugin": "5.5.3",
-    "jest": "29.3.1",
-    "jest-chrome": "0.8.0",
-    "jest-circus": "29.3.1",
-    "jest-dev-server": "6.2.0",
     "jsdom": "22.1.0",
     "postcss": "8.4.30",
     "postcss-loader": "7.3.3",
@@ -326,7 +314,6 @@
     "speed-measure-webpack-plugin": "1.5.0",
     "stream-browserify": "3.0.0",
     "svg-url-loader": "8.0.0",
-    "ts-jest": "29.0.5",
     "ts-node": "10.9.1",
     "ts-unused-exports": "7.0.3",
     "tsconfig-paths-webpack-plugin": "4.0.1",
