+++ src/app/pages/send/send-crypto-asset-form/family/bitcoin/components/bitcoin-send-max-button.tsx
@@ -46,6 +46,7 @@ export function BitcoinSendMaxButton({
     >
       <Box>
         <LeatherButton
+          type="button"
           data-testid={SendCryptoAssetSelectors.SendMaxBtn}
           onClick={() => onSendMax()}
           variant="link"
