+++ src/app/store/transactions/post-conditions.spec.ts
@@ -1,5 +1,5 @@
 import { FungibleConditionCode, parsePrincipalString } from '@stacks/transactions';
-import { HEYSTACK_HEY_TX_REQUEST, HEYSTACK_HEY_TX_REQUEST_DECODED } from '@tests-legacy/mocks';
+import { HEYSTACK_HEY_TX_REQUEST, HEYSTACK_HEY_TX_REQUEST_DECODED } from '@tests/mocks';
 
 import { getPayloadFromToken } from '@shared/utils/requests';
