+++ src/shared/utils/requests.spec.ts
@@ -1,5 +1,5 @@
 import { TransactionTypes } from '@stacks/connect';
-import { HEYSTACK_HEY_TX_REQUEST } from '@tests-legacy/mocks';
+import { HEYSTACK_HEY_TX_REQUEST } from '@tests/mocks';
 
 import { getPayloadFromToken } from './requests';
