--- tests-legacy/global-teardown.ts
@@ -1,6 +1,0 @@
-import { teardown } from 'jest-dev-server';
-
-// ts-unused-exports:disable-next-line
-export default async function globalTeardown() {
-  await teardown();
-}
