--- tests-legacy/integration/asset.selectors.ts
@@ -1,7 +1,0 @@
-// TODO: Revisit when migrate tests
-// export enum AssetSelectors {
-//   AssetSTSW = 'STSW',
-//   AssetDIKO = 'DIKO',
-//   AssetUSDA = 'USDA',
-//   AssetMIA = 'MIA',
-// }
