--- tests-legacy/integration/network.selectors.ts
@@ -1,9 +1,0 @@
-export const NetworkSelectors = {
-  NetworkName: 'network-name',
-  NetworkAddress: 'network-address',
-  NetworkKey: 'network-key',
-  BtnAddNetwork: 'btn-add-network',
-  ErrorText: 'error-text',
-  EmptyAddressError: 'Enter a valid URL',
-  NoNodeFetch: 'Unable to fetch info from node.',
-};
