--- tests-legacy/integration/user-area.selectors.ts
@@ -1,5 +1,0 @@
-export enum UserAreaSelectors {
-  AccountAddress = 'account-address',
-  AccountCopyAddress = 'account-copy-address',
-  AccountCopyBnsAddress = 'account-copy-bns-address',
-}
