--- tests-legacy/page-objects/home.selectors.ts
@@ -1,4 +1,0 @@
-export enum HomePageSelectorsLegacy {
-  BalancesList = 'balances-list',
-  BtnFundAccount = 'btn-fund-account',
-}
