--- tests-legacy/page-objects/wallet.selectors.ts
@@ -1,3 +1,0 @@
-export enum WalletPageSelectors {
-  StatusMessage = 'status-message',
-}
