--- tests-legacy/tsconfig.json
@@ -1,11 +1,0 @@
-{
-  "extends": "../tsconfig.json",
-  "compilerOptions": {
-    "target": "ES2019",
-    "noEmit": true,
-    "rootDir": "../",
-    "esModuleInterop": true,
-    "types": ["jest"]
-  },
-  "include": ["./**/*", "../src/**/*"]
-}
