+++ tests/global-playwright-setup.js
@@ -1,0 +1,8 @@
+import dotenv from 'dotenv';
+
+export default async function globalSetup() {
+  dotenv.config({
+    path: '.env',
+    override: true,
+  });
+}
